+++ CHANGELOG.md
@@ -36,3 +36,12 @@
 * 增加RenderAudioList
 * 增加RenderVideoList
 
+# 1.2.6
+* 增加新的模版指令 BodyTemplate6、ListTemplate3、ListTemplate4
+* 增加音频视频播放信息类 AudioPlayerInfoContent VideoPlayerInfoContent
+* 增加标签类 (AmountTag)数量标签、(AuditionTag)试听标签、(CustomTag)自定义标签、(FreeTag)免费标签、(HotTag)热门标签、(NewTag)最新标签、(PayTag)支付标签、(PurchasedTag)购买标签、(TimeTag)时间标签、(VIPTag)VIP标签
+* 增加端入栈指令 PushStack
+* 增加用户授权指令 AskForPermissionsConsent
+* 增加录音指令 RecordSpeech
+* 增加浏览器访问指令 LaunchBrowser
+* 增加TTS模版 TTSTemplate